refactor: ProfileEditor, NickNameEditor 공용함수화
