+++ app/src/main/java/com/withpeace/withpeace/navigation/NavHost.kt
@@ -44,7 +44,12 @@ fun WithpeaceNavHost(
                 navController.navigateHome()
             },
         )
-        signUpNavGraph(onShowSnackBar = onShowSnackBar)
+        signUpNavGraph(
+            onShowSnackBar = onShowSnackBar,
+            onNavigateToGallery = {
+                navController.navigateToGallery(imageLimit = 1)
+            },
+        )
         registerPostNavGraph(
             onShowSnackBar = onShowSnackBar,
             onCompleteRegisterPost = {},
