+++ core/ui/src/main/res/drawable/ic_default_profile.xml
