+++ core/ui/src/main/res/drawable/ic_editor_pencil.xml
