+++ core/ui/src/main/res/values/strings.xml
@@ -1,15 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <!--  PostTopicUiState   -->
     <string name="free">자유</string>
     <string name="information">정보</string>
     <string name="question">"질문"</string>
     <string name="life">"생활"</string>
     <string name="hobby">"취미"</string>
     <string name="economy">"경제"</string>
+
+    <!--  DateUiState   -->
     <string name="second_format">%1$d초 전</string>
     <string name="minute_format">%1$d분 전</string>
     <string name="hour_format">%1$d시간 전</string>
     <string name="day_format">%1$d일 전</string>
     <string name="years_format">%1$d년 전</string>
 
+    <!-- NicknameEditor   -->
+    <string name="nickname_policy">닉네임은 2~10자의 한글, 영문만 가능합니다.</string>
+    <string name="enter_nickname">닉네임을 입력하세요</string>
+    <string name="nickname_duplicated">중복된 닉네임입니다.</string>
 </resources>
