+++ feature/profileeditor/build.gradle.kts
@@ -7,7 +7,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":core:permission"))
-    implementation(libs.skydoves.landscapist.glide)
-    implementation(libs.skydoves.landscapist.bom)
+
+    implementation(project(":core:ui"))
 }
