+++ feature/profileeditor/src/main/java/com/app/profileeditor/ProfileEditorScreen.kt
@@ -1,12 +1,8 @@
 package com.app.profileeditor
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
@@ -15,50 +11,36 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Divider
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.app.profileeditor.uistate.ProfileEditUiEvent
-import com.app.profileeditor.uistate.ProfileNicknameValidUiState
 import com.app.profileeditor.uistate.ProfileUiModel
-import com.skydoves.landscapist.glide.GlideImage
 import com.withpeace.withpeace.core.designsystem.theme.WithpeaceTheme
 import com.withpeace.withpeace.core.designsystem.ui.WithPeaceBackButtonTopAppBar
-import com.withpeace.withpeace.core.permission.ImagePermissionHelper
+import com.withpeace.withpeace.core.ui.profile.NickNameEditor
+import com.withpeace.withpeace.core.ui.profile.ProfileImageEditor
+import com.withpeace.withpeace.core.ui.profile.ProfileNicknameValidUiState
 import com.withpeace.withpeace.feature.profileeditor.R
-import kotlinx.coroutines.delay
-import kotlin.time.Duration.Companion.seconds
 
 @Composable
 fun ProfileEditorRoute(
@@ -167,19 +149,20 @@ fun ProfileEditorScreen(
                 },
             )
             Spacer(modifier = modifier.height(16.dp))
-            ProfileImage(
+            ProfileImageEditor(
                 profileImage = profileInfo.profileImage,
                 modifier = modifier,
                 onNavigateToGallery = onNavigateToGallery,
+                contentDescription = stringResource(id = R.string.edit_profile),
             )
             Spacer(modifier = modifier.height(24.dp))
             Text(
-                text = stringResource(R.string.nickname_policy),
+                text = stringResource(com.withpeace.withpeace.core.ui.R.string.nickname_policy),
                 style = WithpeaceTheme.typography.caption,
                 color = WithpeaceTheme.colors.SystemGray1,
             )
             Spacer(modifier = modifier.height(16.dp))
-            NickNameTextField(
+            NickNameEditor(
                 nickname = profileInfo.nickname,
                 onNickNameChanged = {
                     onNickNameChanged(it)
@@ -193,141 +176,6 @@ fun ProfileEditorScreen(
         EditCompletedButton(onClick = { onEditCompleted() })
     }
 }
-
-@Composable
-private fun ProfileImage(
-    profileImage: String?,
-    modifier: Modifier,
-    onNavigateToGallery: () -> Unit,
-) {
-    var showDialog by rememberSaveable { mutableStateOf(false) }
-    val context = LocalContext.current
-    val imagePermissionHelper = remember { ImagePermissionHelper(context) }
-    val launcher = imagePermissionHelper.getImageLauncher(
-        onPermissionGranted = onNavigateToGallery,
-        onPermissionDenied = { showDialog = true },
-    )
-    if (showDialog) {
-        imagePermissionHelper.ImagePermissionDialog { showDialog = false }
-    }
-    val imageModifier = modifier
-        .size(120.dp)
-        .clip(CircleShape)
-    Row(
-        modifier = modifier.wrapContentSize(Alignment.Center),
-        horizontalArrangement = Arrangement.Center,
-    ) {
-        Box(
-            modifier.clickable {
-                imagePermissionHelper.onCheckSelfImagePermission(
-                    onPermissionGranted = onNavigateToGallery,
-                    onPermissionDenied = {
-                        imagePermissionHelper.requestPermissionDialog(launcher)
-                    },
-                )
-            },
-        ) {
-            GlideImage(
-                modifier = imageModifier,
-                imageModel = { profileImage },
-                failure = {
-                    Image(
-                        painterResource(id = R.drawable.ic_default_profile),
-                        modifier = imageModifier,
-                        contentDescription = "",
-                    )
-                },
-            )
-            Image(
-                modifier = modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(bottom = 6.dp, end = 6.dp),
-                painter = painterResource(id = R.drawable.ic_editor_pencil),
-                contentDescription = stringResource(id = R.string.edit_profile),
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun NickNameTextField(
-    modifier: Modifier = Modifier,
-    nickname: String,
-    isChanged: Boolean,
-    nicknameValidStatus: ProfileNicknameValidUiState,
-    onNickNameChanged: (String) -> Unit,
-    onKeyBoardTimerEnd: () -> Unit,
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    LaunchedEffect(nickname) {
-        delay(1.seconds)
-        onKeyBoardTimerEnd()
-    }
-
-    Column(
-        modifier = modifier
-            .width(140.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        BasicTextField(
-            value = nickname,
-            onValueChange = {
-                onNickNameChanged(it)
-            },
-            modifier = modifier.fillMaxWidth(),
-            enabled = true,
-            textStyle = WithpeaceTheme.typography.body.copy(textAlign = TextAlign.Center),
-            singleLine = true,
-            maxLines = 1,
-        ) {
-            TextFieldDefaults.DecorationBox(
-                value = nickname,
-                innerTextField = it,
-                enabled = true,
-                singleLine = false,
-                visualTransformation = VisualTransformation.None,
-                placeholder = {
-                    Text(
-                        modifier = modifier.fillMaxWidth(),
-                        textAlign = TextAlign.Center,
-                        text = stringResource(R.string.enter_nickname),
-                        style = WithpeaceTheme.typography.body,
-                        color = WithpeaceTheme.colors.SystemGray2,
-                    )
-                },
-                interactionSource = interactionSource,
-                contentPadding = PaddingValues(0.dp),
-                colors = TextFieldDefaults.colors(
-                    disabledTextColor = Color.Transparent,
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent,
-                    disabledIndicatorColor = Color.Transparent,
-                    unfocusedContainerColor = Color.Transparent,
-                    focusedContainerColor = Color.Transparent,
-                ),
-            )
-        }
-        Divider(
-            color = if (nicknameValidStatus is ProfileNicknameValidUiState.Valid || isChanged.not()) WithpeaceTheme.colors.SystemBlack
-            else WithpeaceTheme.colors.SystemError,
-            modifier = modifier
-                .width(140.dp)
-                .height(1.dp),
-        )
-    }
-    if (nicknameValidStatus !is ProfileNicknameValidUiState.Valid && isChanged) {
-        Text(
-            text = if (nicknameValidStatus is ProfileNicknameValidUiState.InValidDuplicated) stringResource(
-                R.string.nickname_duplicated,
-            ) else stringResource(id = R.string.nickname_policy),
-            style = WithpeaceTheme.typography.caption,
-            color = WithpeaceTheme.colors.SystemError,
-            modifier = modifier.padding(top = 4.dp),
-        )
-    }
-}
-
 
 @Composable
 private fun EditCompletedButton(
@@ -452,4 +300,3 @@ fun ProfileEditorPreview() {
         )
     }
 }
-// uiModel 적용, glide 수정, response 값 전달
