+++ feature/profileeditor/src/main/res/values/strings.xml
@@ -1,11 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="edit_profile">프로필 수정</string>
-    <string name="nickname_policy">닉네임은 2~10자의 한글, 영문만 가능합니다.</string>
     <string name="edit_completed">수정 완료</string>
     <string name="modify_save_request">수정사항이 있습니다.\n저장하시겠습니까?</string>
     <string name="dialog_exit">나가기</string>
     <string name="dialog_save">저장하기</string>
-    <string name="enter_nickname">닉네임을 입력하세요</string>
-    <string name="nickname_duplicated">중복된 닉네임입니다.</string>
 </resources>
