+++ feature/signup/src/main/java/com/withpeace/withpeace/feature/signup/SignUpUiModel.kt
@@ -1,0 +1,6 @@
+package com.withpeace.withpeace.feature.signup
+
+data class SignUpUiModel(
+    val nickname: String,
+    val profileImage: String?,
+)
