+++ feature/signup/src/main/java/com/withpeace/withpeace/feature/signup/navigation/SignUpUiEvent.kt
@@ -1,0 +1,4 @@
+package com.withpeace.withpeace.feature.signup.navigation
+
+interface SignUpUiEvent {
+}
