+++ feature/signup/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="create_profile">프로필을 설정해주세요!</string>
     <string name="nickname_policy">닉네임은 2~10자의 한글, 영문만 가능합니다.</string>
     <string name="nickname_hint">닉네임을 입력하세요</string>
+    <string name="sign_up_completed">가입 완료</string>
+    <string name="set_up_profile_image">프로필 이미지 설정</string>
 </resources>
